🔧 config: update settings (#436)
* 🔧 config: update settings
* 📝 docs: rounding for float32 examples
* 🔧 config(deps): update mpl
* 🔧 config(nox): update config
* 🔧 config(deps): jax constraint
* 🔧 config(deps): bump test deps
* ✅ test: small test adjustments

+++ src/unxt/_interop/unxt_interop_astropy/quantity.py
@@ -253,6 +253,9 @@ def uconvert(u: APYUnits, x: AbstractQuantity, /) -> AbstractQuantity:
     return type_up(x)(**fs)
 
 
+# ============================================================================
+
+
 @dispatch
 def ustrip(u: Any, x: AstropyQuantity) -> Any:
     """Strip the units from the quantity.
@@ -263,8 +266,8 @@ def ustrip(u: Any, x: AstropyQuantity) -> Any:
     >>> import unxt as u
 
     >>> x = apyu.Quantity(1000, "m")
-    >>> u.ustrip(u.unit("m"), x)
-    np.float64(1000.0)
+    >>> float(u.ustrip(u.unit("m"), x))
+    1000.0
 
     """
     return x.to_value(u)
